% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricaCarrots.R
\docType{data}
\name{ricaCarrots}
\alias{ricaCarrots}
\title{The RICA dataset describing the production of carrots (open field) (in quintals) from 2000 to 2015.}
\format{
`ricaCarrots`

A List of 3 objects:
\describe{
  \item{Y}{a vector. The production of carrots (open field) (in quintals) for 598 French farms.}
  \item{X}{a matrix. The 259 covariates describing the same 598 French farms.}
  \item{description}{a matrix. Description of the 259 covariates.}
}
}
\source{
<https://agreste.agriculture.gouv.fr/agreste-web/servicon/I.2/listeTypeServicon/>
}
\usage{
data(ricaCarrots)
}
\description{
A subset of data from the `agreste` French governmental website <https://agreste.agriculture.gouv.fr/agreste-web/servicon/I.2/listeTypeServicon/>.
}
\keyword{datasets}
